RavenDB-23037 Extract shared blittable converter logic into BlittableJsonConverterHelper
Moves serializer-agnostic methods (TryRemoveIdentityProperty, TrySimplifyJson,
GetPropertyType) from the Newtonsoft-specific BlittableJsonConverterBase into a
shared BlittableJsonConverterHelper class. The JObject/JArray/JValue check becomes
a delegate parameter so STJ can provide its own skip logic.
